Add JavaDoc and refactor staff for clean code convention
